intro_cdkchalice/intro_cdkchalice/infrastructure/requirements.txt
aws_cdk.core==1.87.1
aws-cdk.aws-dynamodb==1.87.1
